Handwritten_Math_Expn_Solver/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
opencv-python>=4.5.3
Pillow>=8.3.1
matplotlib>=3.4.3
streamlit>=1.8.0
tensorflow>=2.7.0
sympy>=1.9
scikit-learn>=0.24.2
